requirements.txt
accelerate==0.16.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
faiss-cpu==1.7.4
fastapi==0.98.0
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
html2text==2020.1.16
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.207
langchainplus-sdk==0.0.16
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.5.0
psutil==5.9.5
pydantic==1.10.9
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
yarl==1.9.2
